requirements.txt
streamlit
PyPDF2
pytesseract
pdf2image
Pillow
python-docx
unstructured
spacy
sentence-transformers
torch
numpy
tqdm
